+++ lib/common/colxlate.c
@@ -11,6 +11,7 @@
 #include <stdio.h>
 
 #include <math.h>
+#include <stdbool.h>
 #include <stdlib.h>
 #include <string.h>
 #include <ctype.h>
@@ -254,7 +255,16 @@ int colorxlate(char *str, gvcolor_t * color, color_type_t target_type)
     /* test for rgb value such as: "#ff0000"
        or rgba value such as "#ff000080" */
     a = 255;			/* default alpha channel value=opaque in case not supplied */
-    if (*p == '#' && sscanf(p, "#%2x%2x%2x%2x", &r, &g, &b, &a) >= 3) {
+    bool is_rgb = *p == '#' && sscanf(p, "#%2x%2x%2x%2x", &r, &g, &b, &a) >= 3;
+    if (!is_rgb) { // try 3 letter form
+	is_rgb = *p == '#' && strlen(p) == 4 && sscanf(p, "#%1x%1x%1x", &r, &g, &b) == 3;
+	if (is_rgb) {
+	    r |= r << 4;
+	    g |= g << 4;
+	    b |= b << 4;
+	}
+    }
+    if (is_rgb) {
 	switch (target_type) {
 	case HSVA_DOUBLE:
 	    R = (double) r / 255.0;